+++ Springboot CRUD Project/demo/src/main/java/com/example/demo/ProductRepository.java
@@ -1,10 +1,9 @@
 package com.example.demo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.demo.Product;
 
 //This creates the repository
 
-public interface ProductRepository extends JpaRepository<Product, Integer>{
+public interface ProductRepository extends JpaRepository<Product, Long>{
     
 }
